requirements.txt
appnope==0.1.0
astroid==2.2.5
autopep8==1.4.3
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.2
django-braces==1.13.0
django-cors-headers==2.5.2
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-pipeline==1.6.14
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
ipdb==0.12.2
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.16
jedi==0.13.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.1
parso==0.4.0
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.1
ptyprocess==0.6.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.8
pylint-plugin-utils==0.5
python-decouple==3.1
python-dotenv==0.10.1
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
sendgrid==6.0.4
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sqlparse==0.3.0
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.1
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.11.1
